Fix compiler warnings (#1507)
* Fix compiler warnings

Fixes for compiler warnings.

The majority were mismatched types and some uninitialized variables. Added flags to suppress some deprecation/security warnings.

Remove commented out line

More compiler warning fixes

Added some Intellisense warnings.

* Missed a couple changes

* Update lib-streamlabs-ipc submodule

---------

+++ obs-studio-client/source/advanced-replay-buffer.cpp
@@ -67,7 +67,7 @@ osn::AdvancedReplayBuffer::AdvancedReplayBuffer(const Napi::CallbackInfo &info) 
 {
 	Napi::Env env = info.Env();
 	Napi::HandleScope scope(env);
-	int length = info.Length();
+	size_t length = info.Length();
 
 	if (length <= 0 || !info[0].IsNumber()) {
 		Napi::TypeError::New(env, "Number expected").ThrowAsJavaScriptException();
@@ -89,7 +89,7 @@ Napi::Value osn::AdvancedReplayBuffer::Create(const Napi::CallbackInfo &info)
 	if (!ValidateResponse(info, response))
 		return info.Env().Undefined();
 
-	auto instance = osn::AdvancedReplayBuffer::constructor.New({Napi::Number::New(info.Env(), response[1].value_union.ui64)});
+	auto instance = osn::AdvancedReplayBuffer::constructor.New({Napi::Number::New(info.Env(), static_cast<double>(response[1].value_union.ui64))});
 
 	return instance;
 }
@@ -148,7 +148,7 @@ Napi::Value osn::AdvancedReplayBuffer::GetLegacySettings(const Napi::CallbackInf
 	if (!ValidateResponse(info, response))
 		return info.Env().Undefined();
 
-	auto instance = osn::AdvancedReplayBuffer::constructor.New({Napi::Number::New(info.Env(), response[1].value_union.ui64)});
+	auto instance = osn::AdvancedReplayBuffer::constructor.New({Napi::Number::New(info.Env(), static_cast<double>(response[1].value_union.ui64))});
 
 	return instance;
 }
@@ -183,7 +183,7 @@ Napi::Value osn::AdvancedReplayBuffer::GetStreaming(const Napi::CallbackInfo &in
 	if (!ValidateResponse(info, response))
 		return info.Env().Undefined();
 
-	auto instance = osn::AdvancedStreaming::constructor.New({Napi::Number::New(info.Env(), response[1].value_union.ui64)});
+	auto instance = osn::AdvancedStreaming::constructor.New({Napi::Number::New(info.Env(), static_cast<double>(response[1].value_union.ui64))});
 	return instance;
 }
 
@@ -214,7 +214,7 @@ Napi::Value osn::AdvancedReplayBuffer::GetRecording(const Napi::CallbackInfo &in
 	if (!ValidateResponse(info, response))
 		return info.Env().Undefined();
 
-	auto instance = osn::AdvancedRecording::constructor.New({Napi::Number::New(info.Env(), response[1].value_union.ui64)});
+	auto instance = osn::AdvancedRecording::constructor.New({Napi::Number::New(info.Env(), static_cast<double>(response[1].value_union.ui64))});
 	return instance;
 }
